the kitchen/requirements.txt
Flask==3.1.0
Flask_SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
Flask-CORS==4.0.2
Werkzeug==3.1.3
neon==0.8.0


